Include LICENSE file in package configuration
Added the `LICENSE` file to the `Trumpf.Coparoo.Playwright.Controls.csproj`
project configuration. The file is marked with `Pack="true"` to ensure it
is included in the package during the build process. The `PackagePath=""`
attribute specifies its location in the package, making the license
readily available to consumers of the package.